fix fps calc issue

+++ main.cpp
@@ -123,7 +123,7 @@ int main() {
                     isClient.update(frame, int(timeStamp / 1000));
                     isClient.addText(cv::format("size %d x %d", stFrameInfo.size.width, stFrameInfo.size.height));
                     isClient.addText(cv::format("ts %ld", timeStamp));
-                    isClient.addText(cv::format("fps %3d", int(1000000 / cap->getCurrentInterval())));  // TODO: 大华相机和MindVision返回的值单位不同
+                    isClient.addText(cv::format("fps %3d", 1000000 / cap->getCurrentInterval()));
                     isClient.addText(cv::format("send %2.2f ms", communicator.getCurrentInterval() / 1000.0));
 
                     isClient.clock("run");
@@ -177,12 +177,12 @@ int main() {
                             }
 
                             if (pWindMill->run(frame, pitch, yaw, (double)cv::getTickCount())) {  //有目标，运行风车击打
-                                communicator.send(0.0, 0.0,  SEND_STATUS_WM_AIM, SEND_STATUS_WM_FIND);
+                                communicator.send(0.0, 0.0, SEND_STATUS_WM_AIM, SEND_STATUS_WM_FIND);
                                 isClient.addText(cv::format("send pitch:%0.2f", pitch));
                                 isClient.addText(cv::format("send yaw:%0.2f", yaw));
 
                             } else {  //无目标
-                                communicator.send(0.0, 0.0, 
+                                communicator.send(0.0, 0.0,
                                     SEND_STATUS_WM_AIM, SEND_STATUS_WM_NO);
                                 PRINT_WARN("[windmill] no target find\n");
                             }
